requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
lxml>=4.9.0

# PDF processing
pdfplumber>=0.10.0
pymupdf>=1.23.0

# LLM extraction
openai>=1.12.0

# Database
sqlalchemy>=2.0.0

# Data processing
pandas>=2.1.0
pydantic>=2.5.0

# Web interface
streamlit>=1.29.0
plotly>=5.18.0
altair>=5.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
